addressing comments...
